requirements.txt
pyyaml
configparser
pypinyin
pdfplumber
PyPDF2
python-docx
wikipedia
accelerate
transformers
langchain
sentence_transformers
tiktoken
einops
transformers_stream_generator
pymongo
ninja
fastapi
sse-starlette
uvicorn
gunicorn
supervisor
sseclient-py